Map update - my api key
Basic map.html. Javascript and CSS are in the HTML, rather than in separate files. I'm somewhat unfamililar with your code, so embedding this into the webpage will kinda be up to you. Or I would love to be walked through it with you. The API key used in this code is my personal API key. May want to make a NIMS account if not already made. Can get its own keys for use.